Extract aggregator models (#967)
